requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==3.0.2
django-extensions==2.2.6
djangorestframework==3.11.0
et-xmlfile==1.0.1
httpie==2.0.0
idna==2.8
jdcal==1.4.1
numpy==1.18.1
openpyxl==3.0.3
Pygments==2.5.2
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
